Переделал макет для элемента списка, который показывает картинку сверху и вынес ViewHolder в отдельный класс.
